chore: refactor into bargo lib
